requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pymongo==4.6.0
python-dotenv==1.0.0
aiofiles==23.2.1
aiohttp==3.9.1
motor==3.3.2
dnspython==2.4.2
requests==2.31.0
Pillow==10.1.0
python-magic==0.4.27
asyncio==3.4.3
uvloop==0.19.0
